chore: adjust log levels of the instance endpoints
This patch adjusts several log messages for the HTTP endpoints.

In case of HTTP probe failures, we avoid cluttering the logs by printing an info message, given that this information can still be obtained from pod status.

In case of unmarshal errors, we instead raise the severity from info to warning.

Closes #2958

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>